+++ genomics_data_index/test/integration/storage/io/mutation/test_SequenceFile.py
@@ -7,7 +7,7 @@ from Bio import SeqIO
 
 from genomics_data_index.storage.io.mutation.SequenceFile import SequenceFile
 from genomics_data_index.test.integration import reference_file_5000_snpeff, reference_file_5000_snpeff_2, \
-    reference_file
+    reference_file, reference_file_lzma, reference_file_bzip2
 
 
 def parse_snpeff_config(config_file: Path) -> Dict[str, str]:
@@ -26,6 +26,28 @@ def parse_snpeff_config(config_file: Path) -> Dict[str, str]:
 
 def test_read_fasta_file():
     sequence_file = SequenceFile(reference_file)
+    name, records = sequence_file.parse_sequence_file()
+
+    assert 'genome' == name
+    assert 1 == len(records)
+    record = records[0]
+    assert 'reference' == record.id
+    assert 5180 == len(record)
+
+
+def test_read_fasta_file_lzma():
+    sequence_file = SequenceFile(reference_file_lzma)
+    name, records = sequence_file.parse_sequence_file()
+
+    assert 'genome' == name
+    assert 1 == len(records)
+    record = records[0]
+    assert 'reference' == record.id
+    assert 5180 == len(record)
+
+
+def test_read_fasta_file_bzip2():
+    sequence_file = SequenceFile(reference_file_bzip2)
     name, records = sequence_file.parse_sequence_file()
 
     assert 'genome' == name